chatbot/requirements.txt
gradio==4.19.2
groq==0.4.2
python-dotenv==1.0.1
gTTS==2.5.1
elevenlabs==1.0.0 
speechrecognition==3.10.0
pydub==0.25.1 